Refactored Auth so you can easily change its behavior
